requirements.txt
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
cssselect==1.2.0
Deprecated==1.2.14
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.2
fake-useragent==1.2.1
fastapi==0.101.0
gcloud==0.18.3
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jwcrypto==1.5.0
lxml==4.9.3
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
oauth2client==4.1.3
parse==1.19.1
Pillow==10.0.0
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
pyee==8.2.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
pyppeteer==1.0.2
pyquery==2.0.0
Pyrebase4==4.7.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-jwt==4.0.0
pytweening==1.0.7
PyWinCtl==0.0.43
requests==2.29.0
requests-html==0.10.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
tqdm==4.66.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
w3lib==2.1.2
websockets==10.4
wrapt==1.15.0
zipp==3.16.2
